api/v1/auth/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
httpx==0.28.1
aio-pika==9.5.4
asyncpg==0.30.0